18-contents-summary-importer/pyproject.toml
[project]
name = "18-pdf-summary-importer"
version = "0.1.0"
description = "A script to summarize PDF files from GCS and save the summary to Spanner."
requires-python = ">=3.13"
dependencies = [
    "google-cloud-spanner>=3.59.0",
    "google-genai>=1.49.0",
    "fastapi>=0.121.1",
    "uvicorn[standard]>=0.38.0",
    "ebooklib>=0.20",
    "beautifulsoup4>=4.14.2",
    "google-cloud-storage>=3.5.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.1",
]

[tool.uv.sources.pypi]
url = "https://pypi.org/simple"

